requirements.txt
Flask
Gunicorn
mysql-connector-python
pytz
Flask-Cors
pusher


